refactor generator dir
